feat(ui): wip panneau de recherche & de filtres
